+++ src/ToDo.js
@@ -1,0 +1,9 @@
+class ToDo {
+  constructor(description) {
+    this.description = description;
+    this.completed = false;
+    this.index = 0;
+  }
+}
+
+module.exports = { ToDo };
