--- src/completion.js
@@ -1,6 +1,0 @@
-const createNode = (type, nodeClass) => {
-  const node = document.createElement(type);
-  if (nodeClass) node.className = nodeClass;
-  return node;
-};
-module.exports = createNode;
